digitalocean-builder/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--extra-index-url https://pypi.fury.io/ngjmB-ALDj0b7dH8vBo71WnLEU4SlVxw8/wpengine/

amqp==2.3.2               # via kombu
billiard==3.5.0.5         # via celery
celery[redis]==4.2.1
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==6.7                # via flex, pip-tools
dj-database-url==0.5.0
django-celery-results==1.0.4
django-filter==2.0.0
django-prometheus==1.0.15
django==2.1.4
djangorestframework==3.9.0
flake8==3.6.0             # via wpe-django-logging
flex==6.13.2              # via wpe-django-swagger
httmock==1.2.6            # via wpe-django-auth, wpe-django-swagger
idna==2.8                 # via requests
jsonpointer==1.14         # via flex
kombu==4.2.2.post1        # NOTE: see https://github.com/celery/kombu/issues/974
markdown==3.0.1
mccabe==0.6.1             # via flake8
newrelic==4.8.0.110
pip-tools==3.2.0
prometheus-client==0.5.0  # via django-prometheus
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0        # via flake8
pyflakes==2.0.0           # via flake8
python-dateutil==2.7.5    # via wpe-django-swagger
python-json-logger==0.1.10  # via wpe-django-logging
pytz==2018.7              # via celery, django
pyyaml==3.13              # via flex, wpe-django-auth, wpe-django-swagger
redis==3.0.1              # via celery
requests==2.21.0
rfc3987==1.3.8            # via flex
rollbar==0.14.6
six==1.12.0               # via flex, pip-tools, python-dateutil, rollbar
strict-rfc3339==0.7       # via flex
urllib3==1.24.1           # via requests
uwsgi==2.0.17.1
validate-email==1.3       # via flex
vine==1.1.4               # via amqp
wpe-django-auth==2.0.9
wpe-django-logging==0.1.5
wpe-django-swagger==2.1.20
